api/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
requests
python-dotenv
gunicorn
corsheaders
django-cors-headers
redis==4.5.5
celery[redis]